chore(core): moving squirrel events to it's own handler
Small code tidy.
